Move `ResetPreloadCache` command
