+++ wcfsetup/install/files/lib/command/language/preload/ResetPreloadCache.class.php
@@ -1,25 +1,22 @@
 <?php
 
-namespace wcf\system\language\preload\command;
+namespace wcf\command\language\preload;
 
 use wcf\data\language\Language;
 
 /**
  * Resets the preload cache for the requested language.
  *
- * @author Alexander Ebert
- * @copyright 2001-2022 WoltLab GmbH
- * @license GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
- * @since 6.0
+ * @author      Alexander Ebert
+ * @copyright   2001-2022 WoltLab GmbH
+ * @license     GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
+ * @since       6.2
  */
 final class ResetPreloadCache
 {
-    private readonly Language $language;
-
-    public function __construct(Language $language)
-    {
-        $this->language = $language;
-    }
+    public function __construct(
+        private readonly Language $language
+    ) {}
 
     public function __invoke(): void
     {
