+++ wcfsetup/install/files/lib/system/language/preload/command/ResetPreloadCache.class.php
@@ -7,26 +7,22 @@ use wcf\data\language\Language;
 /**
  * Resets the preload cache for the requested language.
  *
- * @author Alexander Ebert
- * @copyright 2001-2022 WoltLab GmbH
- * @license GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
- * @since 6.0
+ * @author      Alexander Ebert
+ * @copyright   2001-2022 WoltLab GmbH
+ * @license     GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
+ * @since       6.0
+ * @deprecated  6.2 Use `\wcf\command\language\preload\ResetPreloadCache` instead.
  */
 final class ResetPreloadCache
 {
-    private readonly Language $language;
-
-    public function __construct(Language $language)
-    {
-        $this->language = $language;
-    }
+    public function __construct(
+        private readonly Language $language
+    ) {}
 
     public function __invoke(): void
     {
-        // Try to remove the file if it exists.
-        $filename = \WCF_DIR . $this->language->getPreloadCacheFilename();
-        if (\file_exists($filename)) {
-            \unlink($filename);
-        }
+        (new \wcf\command\language\preload\ResetPreloadCache(
+            $this->language
+        ))();
     }
 }
